refactoring surveys
